\name{example.data}
\alias{example.data}
\docType{data}
\title{
Example data
}
\description{
Example data
}
\usage{data("example.data")}
\format{
 A list with 3 elements:
  \describe{
  	  \item{\code{sob}}{the surrogate markers}
    \item{\code{yob}}{the primary outcome}
    \item{\code{aob}}{the treatment indicator}
  	}
}
\examples{
data(example.data)
names(example.data)
}

