\name{Kern.FUN.d}
\alias{Kern.FUN.d}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
helper function
}
\description{
helper function
}
\usage{
Kern.FUN.d(zz, zi, bw)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{zz}{
zz}
  \item{zi}{
zi}
  \item{bw}{
bw}
}
\value{
kernel matrix
}
