\name{Kern.FUN.M}
\alias{Kern.FUN.M}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
helper function
}
\description{
helper function
}
\usage{
Kern.FUN.M(zz, zi, bw)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{zz}{
zz
}
  \item{zi}{
zi
}
  \item{bw}{
bw
}
}
\value{
kernel matrix
}
