\name{resam}
\alias{resam}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Estimates quantities using resampled data}
\description{
Estimates quantities using resampled data
}
\usage{
resam(index, yob, sob, aob, n)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{index}{
index
}
  \item{yob}{
y
}
  \item{sob}{
surrogates
}
  \item{aob}{
treatment
}
  \item{n}{
n
}
}
\value{
Outputs parametric estimate, additative linear estimate, and convex combination estimate
}

