\name{VTM}
\alias{VTM}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Helper function
}
\description{
Helper function; this function creates a matrix that repeats vc, dm times where each row is equal to the vc vector. }
\usage{
VTM(vc, dm)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{vc}{
the vector to repeat.
}
  \item{dm}{
number of rows.
}
}
\value{
a matrix that repeats vc, dm times where each row is equal to the vc vector
}

